Rework Back button behaviour (#369)
* currentFragment could get out of date after Back button

Calling onBackPressed on a non-visible fragment is very bad. Instead of
trying to manually track the active fragment, simply loop through and
find it when we need to know.

* Making Back button behaviour more straightforward.

Rather than trying to maintain the backstack through all the navigation
menu changes (which isn't like behaviour of eg gmail or twitter), aiming
for

* if you're on the journey planner, back leaves the app
* if you're not on the journey planner, back should take you to the journey planner
* unless you're part way through adding a photo, in which case back takes you to the previous step
* or you're on a setting subscreen, in which case back return you to settings
* or you're on the journey planner and planning a route, in which case back removes the most recent waypoint wisp

This change provides all this behaviour, except the setting subscreen.

* Back returns to Settings page after launching an Activity

The Saved Locations and About items on the Settings page launch new
activities. This change ensures we return to the settings page when
those activities are closed with the Back button.

* Tweak SettingsFragment back button behaviour

If on a subscreen, return to the main settings screen

* Updated whats new

Co-authored-by: Oliver Lockwood <oliver.lockwood@cantab.net>