Add PendingIntent.FLAG_IMMUTABLE
This was a breaking change in the enforced move to the new API level. I
don't think there was a build warning for it (although having identified
the problem, the IDE has a squigly under the offending line). Even if
there was I would have missed it in all the deprecation warnings to
which we are now subject.