Resetting the OSMDroid tile cache.
This is relatively drastic, but we've got a number of reports where the
app stops drawing tiles when the phone OS has been upgraded. We don't
currently detect that situation (which can require a cahce reset), but
if we try this on our beta channel we can at least get an idea if it'll
help.

If yes, we can do it properly. Otherwise we can roll back.